Removed redundant tests and refactored the existing test for possible games to use testcases file.
